requirements.txt
advanced_alchemy==1.4.5
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
ansi2html==1.9.2
anyio==4.10.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
bson==0.5.10
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorful==0.5.7
distlib==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
Faker==37.5.3
fastapi==0.116.1
fastapi-utils==0.8.0
filelock==3.18.0
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
litestar==2.16.0
litestar-htmx==0.5.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multipart==1.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.36.0
opentelemetry-exporter-prometheus==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
platformdirs==4.3.8
polyfactory==2.22.1
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==5.9.8
py-spy==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
ray==2.48.0
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rich-click==1.8.9
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.33.1
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
